webkit-efl/LayoutTests/platform/efl-tizen/css3/viewport-percentage-lengths/css3-viewport-percentage-lengths-getStyle-expected.txt

Test for Bug: 27160 - Implement vw/vh/vm (viewport sizes) from CSS 3 Values and Units

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test for vw
PASS style.getPropertyValue("height") is '294px'
PASS style.getPropertyValue("width") is '294px'
PASS style.getPropertyValue("font-size") is '29px'
PASS style.getPropertyValue("line-height") is '39px'
PASS style.getPropertyValue("text-indent") is '2vw'
PASS style.getPropertyValue("margin-left") is '19px'
PASS style.getPropertyValue("margin-right") is '19px'
PASS style.getPropertyValue("margin-top") is '19px'
PASS style.getPropertyValue("margin-bottom") is '19px'
PASS style.getPropertyValue("border-top-left-radius") is '9px'
PASS style.getPropertyValue("border-top-right-radius") is '9px'
PASS style.getPropertyValue("border-bottom-left-radius") is '9px'
PASS style.getPropertyValue("border-bottom-right-radius") is '9px'
PASS style.getPropertyValue("min-height") is '10vw'
PASS style.getPropertyValue("min-width") is '10vw'
PASS style.getPropertyValue("max-height") is '60vw'
PASS style.getPropertyValue("max-width") is '60vw'
PASS style.getPropertyValue("top") is '98px'
PASS style.getPropertyValue("bottom") is '98px'
PASS style.getPropertyValue("left") is '98px'
PASS style.getPropertyValue("right") is '98px'
PASS style.getPropertyValue("padding-left") is '19px'
PASS style.getPropertyValue("padding-right") is '19px'
PASS style.getPropertyValue("padding-top") is '19px'
PASS style.getPropertyValue("padding-bottom") is '19px'

Test for vh
PASS style.getPropertyValue("height") is '220px'
PASS style.getPropertyValue("width") is '220px'
PASS style.getPropertyValue("font-size") is '22px'
PASS style.getPropertyValue("line-height") is '29px'
PASS style.getPropertyValue("text-indent") is '2vh'
PASS style.getPropertyValue("margin-left") is '14px'
PASS style.getPropertyValue("margin-right") is '14px'
PASS style.getPropertyValue("margin-top") is '14px'
PASS style.getPropertyValue("margin-bottom") is '14px'
PASS style.getPropertyValue("border-top-left-radius") is '7px'
PASS style.getPropertyValue("border-top-right-radius") is '7px'
PASS style.getPropertyValue("border-bottom-left-radius") is '7px'
PASS style.getPropertyValue("border-bottom-right-radius") is '7px'
PASS style.getPropertyValue("min-height") is '10vh'
PASS style.getPropertyValue("min-width") is '10vh'
PASS style.getPropertyValue("max-height") is '60vh'
PASS style.getPropertyValue("max-width") is '60vh'
PASS style.getPropertyValue("top") is '73px'
PASS style.getPropertyValue("bottom") is '73px'
PASS style.getPropertyValue("left") is '73px'
PASS style.getPropertyValue("right") is '73px'
PASS style.getPropertyValue("padding-left") is '14px'
PASS style.getPropertyValue("padding-right") is '14px'
PASS style.getPropertyValue("padding-top") is '14px'
PASS style.getPropertyValue("padding-bottom") is '14px'

Test for vmin
PASS style.getPropertyValue("height") is '220px'
PASS style.getPropertyValue("width") is '220px'
PASS style.getPropertyValue("font-size") is '22px'
PASS style.getPropertyValue("line-height") is '29px'
PASS style.getPropertyValue("text-indent") is '2vmin'
PASS style.getPropertyValue("margin-left") is '14px'
PASS style.getPropertyValue("margin-right") is '14px'
PASS style.getPropertyValue("margin-top") is '14px'
PASS style.getPropertyValue("margin-bottom") is '14px'
PASS style.getPropertyValue("border-top-left-radius") is '7px'
PASS style.getPropertyValue("border-top-right-radius") is '7px'
PASS style.getPropertyValue("border-bottom-left-radius") is '7px'
PASS style.getPropertyValue("border-bottom-right-radius") is '7px'
PASS style.getPropertyValue("min-height") is '10vmin'
PASS style.getPropertyValue("min-width") is '10vmin'
PASS style.getPropertyValue("max-height") is '60vmin'
PASS style.getPropertyValue("max-width") is '60vmin'
PASS style.getPropertyValue("top") is '73px'
PASS style.getPropertyValue("bottom") is '73px'
PASS style.getPropertyValue("left") is '73px'
PASS style.getPropertyValue("right") is '73px'
PASS style.getPropertyValue("padding-left") is '14px'
PASS style.getPropertyValue("padding-right") is '14px'
PASS style.getPropertyValue("padding-top") is '14px'
PASS style.getPropertyValue("padding-bottom") is '14px'
PASS successfullyParsed is true

TEST COMPLETE
